add support for encoding/decoding mortal eras
Mortal Eras are currently decoded to a tuple of (period, phase).

When encoding mortal eras, the input value can either be a tuple
(period, phase), which is encoded directly, or a dict.
If a dict is specified, it must contain the field 'period' as well
as either the field 'phase' or 'current'.

If 'phase' is present, the period and phase are encoded directly, as if
a tuple was given with those two values.

If 'current' is present, that is taken as the current block number, and
the phase is calculated for the given period. In this case, the period
will also be converted to the next power of two if it is not already a
power of two. This is to simulate the behavior of the Rust crate's
Era::mortal function which performs the above behavior.